DI customization for tests is extracted as CustomWebApplicationFactory
