+++ tests/MinimalAPI.IntegrationTests/CustomWebApplicationFactory.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.EntityFrameworkCore;
+using MinimalAPI.DataSource.Tables;
+
+namespace MinimalAPI.IntegrationTests
+{
+    public class CustomWebApplicationFactory<TEntryPoint> : WebApplicationFactory<TEntryPoint> 
+        where TEntryPoint : class
+    {
+        protected override void ConfigureWebHost(IWebHostBuilder builder)
+        {
+            builder.ConfigureServices(services =>
+            {
+                var descriptor = services.SingleOrDefault(
+                    d => d.ServiceType == typeof(DbContextOptions<StoreDbContext>));
+
+                if (descriptor != null)
+                {
+                    services.Remove(descriptor);
+                }
+
+                services.AddDbContext<StoreDbContext>(options =>
+                {
+                    options.UseInMemoryDatabase("TestInMemory");
+                });
+            });
+        }
+    }
+}
