+++ tests/MinimalAPI.IntegrationTests/xunit.runner.json
@@ -1,0 +1,3 @@
+{
+	"shadowCopy": false
+}
